requirements.txt
jinja2
pyyaml
watchdog
pillow
